Removed random reactions and diaper check reactions from diapertraining
it was confusing to people and i dont think anyone ended up using it. I also received feedback that you can just see what the state of the diaper is from the image, so w/e.
testing in prod baybe woo

kept the verification check reaction for helping enforcers, though.